Improve data storage

+++ src/main/java/Backups/regalowl/hyperconomy/Backup.java
@@ -5,41 +5,54 @@ import java.util.ArrayList;
 
 public class Backup {
 	Backup() {
-		FileTools ft = new FileTools();
-		ArrayList<String> backupFiles = new ArrayList<String>();
-		backupFiles.add("config.yml");
-		backupFiles.add("displays.yml");
-		backupFiles.add("enchants.yml");
-		backupFiles.add("history.yml");
-		backupFiles.add("items.yml");
-		backupFiles.add("log.yml");
-		backupFiles.add("shops.yml");
-		backupFiles.add("signs.yml");
-		backupFiles.add("categories.yml");
-		backupFiles.add("players.yml");
-		String spath = ft.getJarPath() + File.separator + "plugins" + File.separator + "HyperConomy";
-		String dpath = ft.getJarPath() + File.separator + "plugins" + File.separator + "HyperConomy" + File.separator + "backups";
-		ft.makeFolder(dpath);
-		dpath = dpath + File.separator + ft.getTimeStamp();
-		ft.makeFolder(dpath);
-		for (int i = 0; i < backupFiles.size(); i++) {
-			ft.copyFile(spath + File.separator + backupFiles.get(i), dpath + File.separator + backupFiles.get(i));
-		}
+		HyperConomy hc = HyperConomy.hc;
+		SQLWrite sw = hc.getSQLWrite();
+		int bufferSize = sw.getBufferSize();
 		
-		
-		backupFiles.clear();
-		backupFiles = ft.getFolderContents(spath + File.separator + "Languages");
+		if (bufferSize == 0) {
+			FileTools ft = new FileTools();
+			
+			ArrayList<String> backupFiles = new ArrayList<String>();
+			backupFiles.add("config.yml");
+			backupFiles.add("displays.yml");
+			backupFiles.add("enchants.yml");
+			backupFiles.add("items.yml");
+			backupFiles.add("shops.yml");
+			backupFiles.add("signs.yml");
+			backupFiles.add("categories.yml");
+			if (hc.useMySQL()) {
+				sw.writeData("CREATE TABLE hyperconomy_objects_backup LIKE hyperconomy_objects; INSERT INTO hyperconomy_objects_backup SELECT * FROM hyperconomy_objects");
+				sw.writeData("CREATE TABLE hyperconomy_players_backup LIKE hyperconomy_players; INSERT INTO hyperconomy_players_backup SELECT * FROM hyperconomy_players");
+				sw.writeData("CREATE TABLE hyperconomy_log_backup LIKE hyperconomy_log; INSERT INTO hyperconomy_log_backup SELECT * FROM hyperconomy_log");
+				sw.writeData("CREATE TABLE hyperconomy_audit_log_backup LIKE hyperconomy_audit_log; INSERT INTO hyperconomy_audit_log_backup SELECT * FROM hyperconomy_audit_log");
+			}
+			
+			backupFiles.add("HyperConomy.db");
 
-		spath += File.separator + "Languages";
-		dpath += File.separator + "Languages";
-		ft.makeFolder(dpath);
-		for (int i = 0; i < backupFiles.size(); i++) {
-			ft.copyFile(spath + File.separator + backupFiles.get(i), dpath + File.separator + backupFiles.get(i));
+
+			String spath = ft.getJarPath() + File.separator + "plugins" + File.separator + "HyperConomy";
+			String dpath = ft.getJarPath() + File.separator + "plugins" + File.separator + "HyperConomy" + File.separator + "backups";
+			ft.makeFolder(dpath);
+			dpath = dpath + File.separator + ft.getTimeStamp();
+			ft.makeFolder(dpath);
+			for (int i = 0; i < backupFiles.size(); i++) {
+				if (ft.fileExists(spath + File.separator + backupFiles.get(i))) {
+					ft.copyFile(spath + File.separator + backupFiles.get(i), dpath + File.separator + backupFiles.get(i));
+				}
+			}
+			
+			
+			backupFiles.clear();
+			backupFiles = ft.getFolderContents(spath + File.separator + "Languages");
+	
+			spath += File.separator + "Languages";
+			dpath += File.separator + "Languages";
+			ft.makeFolder(dpath);
+			for (int i = 0; i < backupFiles.size(); i++) {
+				ft.copyFile(spath + File.separator + backupFiles.get(i), dpath + File.separator + backupFiles.get(i));
+			}
+		} else {
+			
 		}
-
-		new HyperPlayersBackup();
-		new HyperHistoryBackup();
-		new HyperLogBackup();
-		new HyperObjectsBackup();
 	}
 }
